Update ex001.c
Exercício:

   Um coração humano bate 1 vez a cada segundo. Faça um programa que dado a idade de uma pessoa um número inteiro, é calculado o número de batimentos do coração dessa pessoa. Considere que um ANO equivale a 365.25 dias. O resultado deverá ser um ponto flutuante com 2 casas decimais.

+++ ex001.c
@@ -1,11 +1,10 @@
 #include <stdio.h>
 
-// (PT-BR) Exercício 1 - Batimento cardíaco
-// (EN-US) Exercise 1 - Heartbeat
+// Exercício 1 - Batimento cardíaco
 int main(void)
 {
-  float age;
-  scanf("%f", &age);
+  int age;
+  scanf("%i", &age);
   float result = age * 60 * 60 * 24 * 365.25;
   printf("O CORACAO BATEU %.2lf VEZES", result);
 }
